+++ src/net/FrameworkDialogs/MessageBox/IMessageBoxShow.cs
@@ -1,0 +1,16 @@
+﻿using System.Windows;
+
+namespace MvvmDialogs.FrameworkDialogs.MessageBox
+{
+    internal interface IMessageBoxShow
+    {
+        MessageBoxResult Show(
+            Window owner,
+            string? messageBoxText,
+            string caption,
+            MessageBoxButton button,
+            MessageBoxImage icon,
+            MessageBoxResult defaultResult,
+            MessageBoxOptions options);
+    }
+}