+++ src/net/FrameworkDialogs/MessageBox/MessageBoxWrapper.cs
@@ -1,14 +1,14 @@
 ﻿using System;
 using System.Windows;
-using WinMessageBox = System.Windows.MessageBox;
 
 namespace MvvmDialogs.FrameworkDialogs.MessageBox
 {
     /// <summary>
-    /// Class wrapping <see cref="WinMessageBox"/>.
+    /// Class wrapping <see cref="System.Windows.MessageBox"/>.
     /// </summary>
     internal sealed class MessageBoxWrapper : IMessageBox
     {
+        private readonly IMessageBoxShow messageBoxShow;
         private readonly MessageBoxSettings settings;
 
         /// <summary>
@@ -16,7 +16,20 @@ namespace MvvmDialogs.FrameworkDialogs.MessageBox
         /// </summary>
         /// <param name="settings">The settings for the folder browser dialog.</param>
         public MessageBoxWrapper(MessageBoxSettings settings)
+            : this(new MessageBoxShow(), settings)
         {
+        }
+
+        /// <summary>
+        /// Initializes a new instance of the <see cref="MessageBoxWrapper"/> class.
+        /// </summary>
+        /// <param name="messageBoxShow">
+        /// The interface responsible for showing the actual message box.
+        /// </param>
+        /// <param name="settings">The settings for the folder browser dialog.</param>
+        public MessageBoxWrapper(IMessageBoxShow messageBoxShow, MessageBoxSettings settings)
+        {
+            this.messageBoxShow = messageBoxShow ?? throw new ArgumentNullException(nameof(messageBoxShow));
             this.settings = settings ?? throw new ArgumentNullException(nameof(settings));
         }
 
@@ -25,13 +38,14 @@ namespace MvvmDialogs.FrameworkDialogs.MessageBox
         {
             if (owner == null) throw new ArgumentNullException(nameof(owner));
 
-            return WinMessageBox.Show(
+            return messageBoxShow.Show(
                 owner,
                 settings.MessageBoxText,
                 settings.Caption,
                 settings.Button,
                 settings.Icon,
-                settings.DefaultResult);
+                settings.DefaultResult,
+                settings.Options);
         }
     }
 }