+++ test/net/FrameworkDialogs/SaveFile/SaveFileDialogSettingsTest.cs
@@ -1,0 +1,27 @@
+using System.Linq;
+using Microsoft.Win32;
+using MvvmDialogs.FrameworkDialogs.Utils;
+using NUnit.Framework;
+
+namespace MvvmDialogs.FrameworkDialogs.SaveFile
+{
+    [TestFixture]
+    public class SaveFileDialogSettingsTest
+    {
+        [Test]
+        public void NativeDialogSettingsParity()
+        {
+            // Arrange
+            var settingsPropertyNames = string.Join(
+                ", ",
+                DialogSettings.GetPropertyNames(typeof(SaveFileDialogSettings)));
+
+            var dialogPropertyNames = string.Join(
+                ", ",
+                DialogSettings.GetPropertyNames(typeof(SaveFileDialog)).Except(DialogSettings.ExcludedPropertyNames));
+
+            // Assert
+            Assert.That(settingsPropertyNames, Is.EqualTo(dialogPropertyNames));
+        }
+    }
+}