+++ CHANGELOG.md
@@ -6,16 +6,42 @@ This project adheres to [Semantic Versioning](http://semver.org/) and is followi
 
 ## Unreleased
 
+### :zap: Added
+
+- [#91](https://github.com/FantasticFiasco/mvvm-dialogs/issues/91) Added the following settings for parity with the native Windows dialogs (proposed by [@Kimmerest](https://github.com/Kimmerest))
+    - `FolderBrowserDialogSettings`
+        - `RootFolder` - The root folder where the browsing starts from
+    - `MessageBoxSettings`
+        - `Options` - A value object that specifies the options
+    - `OpenFileDialogSettings`
+        - `CustomPlaces` - The list of custom places for file dialog boxes
+        - `DereferenceLinks` - A value indicating whether a file dialog returns either the location of the file referenced by a shortcut or the location of the shortcut file (.lnk)
+        - `ReadOnlyChecked` - A value indicating whether the read-only check box displayed by the open file dialog is selected
+        - `SafeFileName` - A string that only contains the file name for the selected file
+        - `SafeFileNames` - An array that contains one safe file name for each selected file
+        - `ShowReadOnly` - A value indicating whether the open file dialog contains a read-only check box
+        - `ValidateNames` - A value indicating whether the dialog accepts only valid Win32 file names
+    - `SaveFileDialogSettings`
+        - `CustomPlaces` - The list of custom places for file dialog boxes
+        - `DereferenceLinks` - A value indicating whether a file dialog returns either the location of the file referenced by a shortcut or the location of the shortcut file (.lnk)
+        - `SafeFileName` - A string that only contains the file name for the selected file
+        - `SafeFileNames` - An array that contains one safe file name for each selected file
+        - `ValidateNames` - A value indicating whether the dialog accepts only valid Win32 file names
+
+### :syringe: Changed
+
+- [BREAKING CHANGE] The default value of `SaveFileDialogSettings.CheckFileExists` has changed from `true` to `false`, aligning it with the default value of the native Windows `SaveFileDialog`
+
 ## 6.0.0 - 2019-11-02
 
 ### :zap: Added
 
-- [#55](https://github.com/FantasticFiasco/mvvm-dialogs/issues/55) - Support for .NET Core 3.0
-- [#55](https://github.com/FantasticFiasco/mvvm-dialogs/issues/55) - NuGet package supporting .NET 4.5.2, .NET 4.6.2 and .NET 4.7.2
+- [#55](https://github.com/FantasticFiasco/mvvm-dialogs/issues/55) Support for .NET Core 3.0 (proposed by [@virzak](https://github.com/virzak))
+- [#55](https://github.com/FantasticFiasco/mvvm-dialogs/issues/55) NuGet package supporting .NET 4.5.2, .NET 4.6.2 and .NET 4.7.2
 
 ### :dizzy: Changed
 
-- Adaptation to [nullable references in C# 8.0](https://docs.microsoft.com/en-us/dotnet/csharp/nullable-references) introduced breaking changes to the API
+- [BREAKING CHANGE] Adaptation to [nullable references in C# 8.0](https://docs.microsoft.com/en-us/dotnet/csharp/nullable-references) introduced breaking changes to the API
 
 ## 5.3.0 - 2019-04-29
 
@@ -27,13 +53,13 @@ This project adheres to [Semantic Versioning](http://semver.org/) and is followi
 
 ### :syringe: Fixed
 
-- `Logger.Writer` does not throw exception when being set to `null` (contribution by [@MathewSachin](https://github.com/MathewSachin))
+- `Logger.Writer` does not throw exception when being set to `null` (contributed by [@MathewSachin](https://github.com/MathewSachin))
 
 ## 5.2.1 - 2018-10-30
 
 ### :syringe: Fixed
 
-- [#49](https://github.com/FantasticFiasco/mvvm-dialogs/issues/49) - `FilterIndex` wasn't updated on closed Open/Save File dialog
+- [#49](https://github.com/FantasticFiasco/mvvm-dialogs/issues/49) `FilterIndex` wasn't updated on closed Open/Save File dialog (discovered by [@ericdc66](https://github.com/ericdc66))
 
 ## 5.2.0 - 2018-10-11
 
@@ -51,7 +77,7 @@ This project adheres to [Semantic Versioning](http://semver.org/) and is followi
 
 ### :zap: Added
 
-- [#30](https://github.com/FantasticFiasco/mvvm-dialogs/issues/30) Support for customizing the following Windows dialogs (contribution by [@pdinnissen](https://github.com/pdinnissen)):
+- [#30](https://github.com/FantasticFiasco/mvvm-dialogs/issues/30) Support for customizing the following Windows dialogs (contributed by [@pdinnissen](https://github.com/pdinnissen)):
     - Message box
     - Open file dialog
     - Save file dialog
@@ -59,7 +85,7 @@ This project adheres to [Semantic Versioning](http://semver.org/) and is followi
 
 ### :skull: Removed
 
-- Reduce the constructors overloads in `DialogService` into:
+- [BREAKING CHANGE] Reduce the constructors overloads in `DialogService` into:
     - `ctor()` - Default constructor that takes no arguments
     - `ctor(IDialogFactory?, IDialogTypeLocator?, IFrameworkDialogFactory?)` - Constructor that allows full customization
 
@@ -67,7 +93,7 @@ This project adheres to [Semantic Versioning](http://semver.org/) and is followi
 
 ### :zap: Added
 
-- Option to set filter index when opening or saving a file (contribution by [@gigios](https://github.com/gigios))
+- Option to set filter index when opening or saving a file (contributed by [@gigios](https://github.com/gigios))
 
 ## 4.0.10 - 2017-04-20
 
@@ -79,27 +105,27 @@ This project adheres to [Semantic Versioning](http://semver.org/) and is followi
 
 ### :zap: Added
 
-- [#14](https://github.com/FantasticFiasco/mvvm-dialogs/issues/14) - Support for showing custom WPF dialogs by implementing `IWindow`
+- [#14](https://github.com/FantasticFiasco/mvvm-dialogs/issues/14) [BREAKING CHANGE] Support for showing custom WPF dialogs by implementing `IWindow` (proposed by [@wakuflair](https://github.com/wakuflair))
 - Support for showing custom UWP content dialogs by implementing `IContentDialog`
 
 ## 3.0.0 - 2016-09-21
 
 ### :dizzy: Changed
 
-- Updated the constructors of `DialogService`, making the class more friendly to IoC containers
+- [BREAKING CHANGE] Updated the constructors of `DialogService`, making the class more friendly to IoC containers
 
 ## 2.0.1 - 2016-05-18
 
 ### :zap: Added
 
-- [#8](https://github.com/FantasticFiasco/mvvm-dialogs/issues/8) - Support for Universal Windows Platform (UWP)
+- [#8](https://github.com/FantasticFiasco/mvvm-dialogs/issues/8) Support for Universal Windows Platform (UWP) (proposed by [@MrWolfPST](https://github.com/MrWolfPST))
 - Add `ViewNotRegisteredException` that is thrown by `IDialogService` when view is unregistered
 
 ## 1.2.16 - 2016-01-13
 
 ### :syringe: Fixed
 
-- [#2](https://github.com/FantasticFiasco/mvvm-dialogs/issues/2) - Fixed issue where exception was thrown when unregistering a view
+- [#2](https://github.com/FantasticFiasco/mvvm-dialogs/issues/2) Fixed issue where exception was thrown when unregistering a view (discovered by [@Andy360](https://github.com/Andy360))
 
 ## 1.2.0 - 2015-12-14
 
@@ -111,7 +137,7 @@ This project adheres to [Semantic Versioning](http://semver.org/) and is followi
 
 ### :syringe: Fixed
 
-- [#3](https://github.com/FantasticFiasco/mvvm-dialogs/issues/3) - Fixed issue where exception no longer is thrown when the property changed event for `IModalDialogViewModel.DialogResult` is sent twice
+- [#3](https://github.com/FantasticFiasco/mvvm-dialogs/issues/3) Fixed issue where exception no longer is thrown when the property changed event for `IModalDialogViewModel.DialogResult` is sent twice (discovered by [@igitur](https://github.com/igitur))
 
 ## 1.1.3 - 2015-07-07